Refactor: drawPattern 로직 분리함
